requirements.txt
asgiref==3.2.9
beautifulsoup4==4.9.1
bleach==3.1.5
certifi==2020.6.20
chardet==3.0.4
Django==3.0.8
django-ckeditor==5.9.0
django-comments-tree==0.2.0
django-contrib-comments==1.9.2
django-crispy-forms==1.9.1
django-js-asset==1.2.2
django-markupfield==2.0.0
django-mptt==0.11.0
django-treebeard==4.3.1
django-widget-tweaks==1.4.8
djangorestframework==3.11.0
docutils==0.16
draftjs-exporter==4.1.0
filebrowser-safe==0.5.0
future==0.18.2
grappelli-safe==0.5.2
idna==2.10
importlib-metadata==1.7.0
Markdown==3.2.2
oauthlib==3.1.0
packaging==20.4
Pillow==7.1.2
pyparsing==2.4.7
pytz==2020.1
requests==2.24.0
requests-oauthlib==1.3.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
tzlocal==2.1
urllib3==1.25.10
webencodings==0.5.1
zipp==3.1.0
